baseval() method in IntView and all derivedViews.
Various propagators may point to the same variable implementation with
different derived views. It is thus impossible to communicate information
between propagators about a pair (var,val) without a conversion to
the base value of a view. baseval() is a way of converting the derived
value to the base value before calling the set() method of the interface
CBS.