cbs() method now returns a boolean. CBS argument is now a pointer.
The cbs() method in the base class returns false, meaning it is not
overloaded. All propagators that overloads cbs() must return true. Also,
they must check if CBS* is null. If it is, they must return true
without doing any calculations. This behavior can be use for branching
in the status() method to check if there exists any propagator that
overloaded the cbs() method.